move pareto.py back to utils
we dont use Box anymore, circular import is avoided
